Harmonize how covariance matrices are initialized
